% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lift_chart.R
\name{calibration_curve}
\alias{calibration_curve}
\title{Plot the calibration curve for the binary predictions.}
\usage{
calibration_curve(table, bin_number = 10)
}
\arguments{
\item{table}{The data.frame with two columns named by \code{y} and \code{yhat}.}

\item{bin_number}{default 10. The binning set number.}
}
\description{
Plot the calibration curve for the binary predictions.
}
\examples{
\dontrun{calibration_curve(add2evaluation::df, bin_number = 10)}
}
