% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lift_chart.R
\name{cum_gains_chart}
\alias{cum_gains_chart}
\title{Plot the cumulative gains chart.}
\usage{
cum_gains_chart(table)
}
\arguments{
\item{table}{The data.frame with two columns named by \code{y} and \code{yhat}.}
}
\description{
Plot the cumulative gains chart.
}
\examples{
\dontrun{cum_gains_chart(add2evaluation::df)}
}
