% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lift_chart.R
\name{cum_lift_chart}
\alias{cum_lift_chart}
\title{Plot the cumulative lift chart.}
\usage{
cum_lift_chart(table)
}
\arguments{
\item{table}{The data.frame with two columns named by \code{y} and \code{yhat}.}
}
\description{
Plot the cumulative lift chart.
}
