% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataset.R
\docType{data}
\name{dataset}
\alias{dataset}
\title{A sample data from the \url{www.dcjingsai.com}}
\format{A data frame with 183093 rows and 4 variables:
\describe{
  \item{short_name}{the id of machines}
  \item{t}{observation time}
  \item{p}{target}
  \item{phat}{predictions}
}}
\source{
\url{https://github.com/JiaxiangBU/phv}
}
\usage{
dataset
}
\description{
A dataset containing four columns, \code{short_name}, \code{t}
\code{p}, and \code{phat}.
}
\keyword{datasets}
