% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lift_chart.R
\name{decile_chart}
\alias{decile_chart}
\title{Plot the decile chart for the binary predictions.}
\usage{
decile_chart(table, bin_number = 10)
}
\arguments{
\item{table}{The data.frame with two columns named by \code{y} and \code{yhat}.}

\item{bin_number}{default 10. The binning set number.}
}
\description{
Plot the decile chart for the binary predictions.
}
\examples{
\dontrun{decile_chart(add2evaluation::df, bin_number = 10)}
}
