% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/df.R
\docType{data}
\name{df}
\alias{df}
\title{A sample with the target and predictions}
\format{A data frame with 500 rows and 2 variables:
\describe{
  \item{yhat}{predictions}
  \item{y}{target}
}}
\source{
package \code{yardstick::two_class_example}
}
\usage{
df
}
\description{
A dataset containing two columns, \code{yhat} and \code{y}
}
\keyword{datasets}
