% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/groom_model.R
\name{groom_model}
\alias{groom_model}
\title{Get model details by \pkg{broom}}
\usage{
groom_model(model)
}
\arguments{
\item{model}{The model object supported by \pkg{broom}}
}
\description{
Get model details by \pkg{broom}
}
