% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lift_chart.R
\name{lift_chart}
\alias{lift_chart}
\title{Plot the lift chart for the binary predictions.}
\usage{
lift_chart(table, bin_number = 10)
}
\arguments{
\item{table}{The data.frame with two columns named by \code{y} and \code{yhat}.}

\item{bin_number}{default 10. The binning set number.}
}
\description{
Plot the lift chart for the binary predictions.
}
