% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lift_chart.R
\name{lift_table}
\alias{lift_table}
\title{Output the dataframe table of lift chart.}
\usage{
lift_table(table, bin_number = 10)
}
\arguments{
\item{table}{The data.frame with two columns named by \code{y} and \code{yhat}.}

\item{bin_number}{default 10. The binning set number.}
}
\description{
Output the dataframe table of lift chart.
}
