% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pca_inbar.R
\name{pca_inbar}
\alias{pca_inbar}
\title{Run PCA model for a dataset and build a plot.}
\usage{
pca_inbar(data)
}
\arguments{
\item{data}{data.frame prepared for running PCA model.}
}
\description{
Run PCA model for a dataset and build a plot.
}
