% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phv_metric.R
\name{phv_metric}
\alias{phv_metric}
\title{Build the evaluation metrics for the competition phv.}
\usage{
phv_metric(id, t, y, yhat)
}
\arguments{
\item{id}{id.}

\item{t}{The time column.}

\item{y}{Target.}

\item{yhat}{Predictions.}
}
\description{
Build the evaluation metrics for the competition phv.
}
